requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
feedparser==6.0.10
idna==3.4
mutagen==1.46.0
pickleDB==0.9.2
pyaes==1.6.1
Pyrogram==2.0.104
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
sgmllib3k==1.0.0
TgCrypto==1.2.5
urllib3==1.26.15
